requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
watchdog==6.0.0
pypdf==5.4.0
pillow==12.1.1
pypdfium2==5.6.0
winrt-Windows.Foundation==3.2.1
winrt-Windows.Foundation.Collections==3.2.1
winrt-Windows.Graphics.Imaging==3.2.1
winrt-Windows.Media.Ocr==3.2.1
winrt-Windows.Storage.Streams==3.2.1
